# app/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Statistics service for finance-manager application",
        "title": "Stats-service API",
        "contact": {
            "name": "Anton",
            "email": "ap363402@gmail.com"
        },
        "license": {
            "name": "Apache 2.0"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/api/operations": {
            "get": {
                "description": "Retrieves a list of operations with support for filtering and sorting.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Operations"
                ],
                "summary": "Get operations",
                "operationId": "get-operations",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User UUID",
                        "name": "user_uuid",
                        "in": "path"
                    },
                    {
                        "type": "string",
                        "description": "Category name",
                        "name": "category_name",
                        "in": "path"
                    },
                    {
                        "type": "string",
                        "description": "Category type",
                        "name": "query",
                        "in": "path"
                    },
                    {
                        "type": "string",
                        "description": "Category ID",
                        "name": "category_id",
                        "in": "path"
                    },
                    {
                        "type": "string",
                        "description": "Description",
                        "name": "description",
                        "in": "path"
                    },
                    {
                        "type": "string",
                        "description": "Money sum (supports operators for numbers: eq, neq, lt, lte, gt, gte)",
                        "name": "money_sum",
                        "in": "path"
                    },
                    {
                        "type": "string",
                        "description": "Date and time of operation (supports formats: yyyy-mm-dd, yyyy-mm-dd:yyyy-mm-dd)",
                        "name": "date_time",
                        "in": "path"
                    },
                    {
                        "type": "string",
                        "description": "Field to sort by (money_sum, date_time, description)",
                        "name": "sort_by",
                        "in": "path"
                    },
                    {
                        "type": "string",
                        "description": "Sort order (asc, desc)",
                        "name": "sort_order",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of operations",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}